+++ pages/_slug.vue
@@ -1,16 +1,15 @@
 <template>
   <v-content>
     <v-container>
-      <v-layout row wrap align-center>
-        <v-flex xs12 md4>
+      <v-layout row wrap align-start>
+        <v-flex xs12 md3>
           <div class="text-xs-center">
-            <v-avatar size="125px">
+            <v-avatar size="130px">
               <img
-                class="img-circle elevation-7 mb-1"
-                :src="this.$helper.avatar(user)"
-              >
+                class="elevation-7 mb-1"
+                :src="this.$helper.avatar(user)">
             </v-avatar>
-            <div class="headline iranblack">{{this.$persianJS.arabicChar(user.name)}}</div>
+            <div class="headline iranblack light-blue--text text--darken-3">{{this.$persianJS.arabicChar(user.name)}}</div>
             <div class="subheading text-xs-center grey--text pt-1 pb-3">{{this.$persianJS.englishNumber(user.std_numbers)}}</div>
             <v-layout justify-space-between>
               <a href="javascript:;" class="body-2">Home</a>
@@ -22,7 +21,7 @@
         </v-flex>
         <v-flex xs12 md5 offset-md2>
           <div v-for="post in user.posts" :key="post._id">
-            <v-card class="my-3" hover>
+            <v-card class="my-3" flat>
               <v-card-media
                 v-if="post.image || post.image !== ''"
                 class="white--text"
@@ -72,7 +71,14 @@
           }).catch(e => {
             context.error({ statusCode: 404, message: 'کاربر مورد نظر یافت می‌نشود...' })
           })
-      }
+      },
+      // fetch ({ store, redirect }) {
+      //   if (store.$auth.loggedIn) {
+      //     return redirect('/profile')
+      //   }
+      //   else
+      //     return redirect('/login')
+      // },
     }
 </script>
